Finish per-group energy buffers in mixed precision; fix isolated-bonded posq download
Completes the mixed-precision energy-buffer conversion across the remaining forces
that expose a per-group energy buffer (isolated bonded, isolated site, GBSA grid)
so all six forces NUTS gathers from are consistent, and makes the NUTS
nutsGatherForceEnergies reader cast the buffers as mixed* (was float*) to match.

Also fixes a correctness bug in CudaIsolatedBondedKernels: computeInternalEntropy
and computeInternalForceConstants downloaded cu.getPosq() into vector<float4>,
which corrupts/throws in a double-precision context (posq is double4). Added a
precision-aware downloadPositionsVec3 helper.

The deterministic fixed-point main-energy accumulators are unchanged. HMC does not
read these per-group buffers, so it is unaffected. Existing suites pass (36 pytest,
6 aux Hessian).

+++ platforms/cuda/src/CudaGBSAGridForceKernels.h
@@ -162,8 +162,8 @@ private:
     CUmodule generationModule;
 
     // Host-side results cache
-    mutable std::vector<float> groupEnergiesHost;
-    mutable std::vector<float> groupLigandEnergiesHost;
+    mutable std::vector<double> groupEnergiesHost;
+    mutable std::vector<double> groupLigandEnergiesHost;
     mutable std::vector<std::vector<float>> groupBornRadiiHost;
 
     // Hessian support
